+++ inst/examples/virtual-select/default/app.R
@@ -18,6 +18,13 @@ ui <- fluidPage(
         label = "Multiple select:",
         choices = setNames(month.abb, month.name),
         multiple = TRUE
+      ),
+      virtualSelectInput(
+        inputId = "onclose",
+        label = "Update value on close:",
+        choices = setNames(month.abb, month.name),
+        multiple = TRUE,
+        updateOn = "close"
       )
     ),
     column(
@@ -32,7 +39,14 @@ ui <- fluidPage(
       tags$b("Multiple select :"),
       verbatimTextOutput("res_multiple"),
       tags$b("Is virtual select open ?"),
-      verbatimTextOutput(outputId = "res_multiple_open")
+      verbatimTextOutput(outputId = "res_multiple_open"),
+
+      tags$br(),
+
+      tags$b("Update on close :"),
+      verbatimTextOutput("res_onclose"),
+      tags$b("Is virtual select open ?"),
+      verbatimTextOutput(outputId = "res_onclose_open")
     )
   )
 
@@ -46,6 +60,9 @@ server <- function(input, output, session) {
 
   output$res_multiple <- renderPrint(input$multiple)
   output$res_multiple_open <- renderPrint(input$multiple_open)
+
+  output$res_onclose <- renderPrint(input$onclose)
+  output$res_onclose_open <- renderPrint(input$onclose_open)
 
 }